+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # CHANGELOG
 
+## (2022-02-03)
+- added aws-classic-py-vpc-quickstart
+
+## (2022-02-02)
+- added aws-classic-ts-vpc-quickstart
+
 ## (2022-01-26)
 - added azure-ts-synapse-workspace